+++ Changes.md
@@ -19,6 +19,7 @@ Improvements
   - Moved shaders to `3Delight/Shader` menu and removed outdated shaders from the menu.
   - Shaders (including light shaders) are only loaded from the `osl` subdirectory of the 3Delight installation.
   - Primitive variables named `uv` are now automatically renamed `st` for compatibility with the `uvCoord` shader's expectation.
+  - Added a default `uvCoord` shader during internal shader network preprocessing to shader parameters that do not have an input connection.
 
 Fixes
 -----